src_phase_2/requirements.txt
albumentations==1.2.0
altair==4.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
catalogue==2.0.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
dill==0.3.5.1
docker-pycreds==0.4.0
einops==0.4.1
entrypoints==0.4
executing==0.8.3
fastcore==1.4.5
fonttools==4.33.3
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imageio==2.19.3
imgaug==0.4.0
importlib-resources==5.8.0
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonlines==3.0.0
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
kiwisolver==1.4.3
loguru==0.6.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
networkx==2.8.4
numpy==1.23.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
qudida==0.0.4
requests==2.28.0
rich==12.4.4
scikit-image==0.18.3
scikit-learn==1.1.1
scipy==1.8.1
sentry-sdk==1.6.0
setproctitle==1.2.3
Shapely==1.8.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
srsly==2.4.3
stack-data==0.3.0
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.5.4
toolz==0.11.2
torch==1.10.1+cu111
torch-snippets==0.496
torchaudio==0.10.1+rocm4.1
torchsummary==1.5.1
torchvision==0.11.2+cu111
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing==3.7.4.3
typing_extensions==4.2.0
urllib3==1.26.9
wandb==0.12.19
wasabi==0.9.1
wcwidth==0.2.5
zipp==3.8.0
